fixed the  resolution glitch
the  resolution glitch is the glitch where you die and thew resolutuion
changes, i didnt fix it in the best way since there is an object that
makes it all change resolution that i didnt find, since the code could
be hidden absolutely anywhere in the game i decided to make a ciode that
covers the glitch, the code is located in objPlayer step and it is
unititled. again, i dont know where the other code is located so that is
the best solution for now, the only way we can find the other code is by
searching throught all objects. unless you yourself made the code and
know where it is located